+++ src/interfaces/pagination.ts
